refactor: share extractYoutubeId too, not just the scan
Sharing the scan took duplication from 22.1% to 4.2%, still over the 3%
gate. The remainder was the wrapper around it: 20 lines at 90%
duplicated between `surveys` and `apps/web`, which is the same function
with the same four patterns in the same order.

It moves next to the scan it calls, and all three modules import it.
`extractYoutubeId` had no callers outside its own module and their
tests, so this is not an API change for anything.

survey-ui's copy used named capture groups (`(?<videoId>…)`) where the
other two used positional. They resolve to the same string, so the
shared version keeps the positional form — confirmed rather than
assumed: survey-ui is identical to base across 50k inputs after the
change, as are the other two.

Everything else about the three modules stays separate. They are
parallel by design, one per build target; only the two pieces whose
semantics are subtle now have a single home.
